Skeleton of the extension point implementation
