+++ src/main/java/org/jenkinsci/plugins/securityinspector/model/SecurityInspectorReport.java
@@ -22,7 +22,7 @@
  * THE SOFTWARE.
  */
 
-package org.jenkinsci.plugins.securityinspector;
+package org.jenkinsci.plugins.securityinspector.model;
 
 import java.util.Set;
 import java.util.Collection;
